+++ changelog.md
@@ -5,6 +5,7 @@
   - Changed command usage
   - Updated README and added a warning to use the [Anti-VPN-Service](github.com/xpdustry/Anti-VPN-Service) plugin for IP and subnet filtering
   - Added colors when displaying plugin config
+  - Added command to client-side, for admin players
 
 #### 1.5:
   - Fixed compilation issue with Mindustry sources and changed to Toxopid